refactor(redux): redesigned redux file structure according to edits
