+++ src/components/Card/Card.tsx
@@ -1,31 +1,22 @@
 import { FunctionComponent, useEffect, useState } from "react";
-import { ICommentsStorage } from "../../interfaces/interfaces";
+import { useSelector } from "react-redux";
+import commentImg from "../../assets/comment.svg";
+import { selectors } from "../../redux/ducks";
 import { ICardProps } from "./interfaces";
-import { Content, Comments, CommentsCount } from "./styles";
+import { Content, Comments, CommentsCount, CommentsImg } from "./styles";
 
-const Card: FunctionComponent<ICardProps> = ({
-  title,
-  id,
-  // getComments,
-  // comments,
-}) => {
-  // const [localComments, setLocalComments] = useState<ICommentsStorage[]>(
-  //   getComments(id, comments)
-  // );
-
-  // useEffect(() => {
-  //   setLocalComments(getComments(id, comments));
-  // }, [comments]);
+const Card: FunctionComponent<ICardProps> = ({ title, id }) => {
+  const localComments = useSelector(selectors.comments.getCommentsById(id));
 
   return (
     <Content>
       {title}
-      {/* {localComments.length > 0 ? (
+      {localComments.length > 0 ? (
         <Comments>
-          <img src={"img/comment.svg"} />
+          <CommentsImg src={commentImg} />
           <CommentsCount>{localComments.length}</CommentsCount>
         </Comments>
-      ) : null} */}
+      ) : null}
     </Content>
   );
 };
