+++ src/redux/ducks/Auth/selectors.ts
@@ -1,3 +1,3 @@
-import { rootState } from "../store";
+import { rootState } from "../../store";
 
 export const getAuthorName = (state: rootState) => state.authorNames;
