+++ src/redux/ducks/Card/selectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from "reselect";
-import { rootState } from "../store";
+import { rootState } from "../../store";
 
 export const selectCard = (state: rootState) => state.cards;
