+++ src/redux/ducks/Collumns/slice.ts
@@ -1,4 +1,4 @@
-import { ICollumnStorage } from "../../interfaces/interfaces";
+import { ICollumnStorage } from "../../../interfaces/interfaces";
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
 const initialState: ICollumnStorage[] = [
