+++ src/redux/ducks/Comments/slice.ts
@@ -1,4 +1,4 @@
-import { ICommentsStorage } from "../../interfaces/interfaces";
+import { ICommentsStorage } from "../../../interfaces/interfaces";
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
 const initialState: ICommentsStorage[] = [];
